add Controller layer to project

+++ src/main/java/com/practice/intern/Book/Management/API/demo/model/Book.java
@@ -16,11 +16,11 @@ public class Book {
     private double price;
 
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
